+++ CHANGELOG.md
@@ -18,6 +18,7 @@ A preview of major changes can be found in the Wiki ([Latest Changes](https://gi
 - Handle new setups where `version.json` doesn't exist
 - Bug where `max_rounds` would overwrite results from previous legs
 - Issue with adding players to tournament when no more players exist
+- Correctly show Max Rounds dialog if player busts on last visit
 
 ## [2.9.0] - 2025-04-06
 #### Feature